swipe_to_delete history ui
